+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Ability to export/import jobs.
+- Periodic email alerts of new crashes.
 - Button to purge crash records from database.
 
 ### Changed